pyproject.toml
# pyproject.toml
[project]
name = "garage-inventory"
version = "2.0.0"
description = "A Flask-based web application for managing storage boxes using QR codes"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "flask-sqlalchemy>=3.1.0",
    "flask-login>=0.6.0",
    "flask-wtf>=1.2.0",
    "flask-mail>=0.10.0",
    "flask-admin>=1.6.0",
    "werkzeug>=3.0.0",
    "itsdangerous>=2.1.0",
    "python-dotenv>=1.0.0",
    "qrcode[pil]>=7.4.0",
    "pillow>=10.0.0",
    "boto3>=1.34.0",
    "gunicorn>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-flask>=1.3.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.4.0",
    "moto>=5.0.0",
]

# ENTRY POINTS - This is what you asked about!
# After running 'uv sync', you can use these commands:
[project.scripts]
# Development: just run 'garage-dev' from terminal
garage-dev = "garage.__main__:main"

# For production with gunicorn, you'd run:
#   gunicorn "garage:create_app()"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Tell Python where to find the 'garage' package
[tool.hatch.build.targets.wheel]
packages = ["src/garage"]

# Ruff linter configuration
[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"
